add back support for native queries
